pyproject.toml
[project]
name = "socker-strategy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "openai>=1.0.0",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.27.0",
]
